PUT changed to POST for sendBeacon to work + some other fixes

+++ backend/Contracts/ShoppingLists/Update/ShoppingSublistUpdateDto.cs
@@ -4,6 +4,6 @@ namespace Cookbook.Contracts.ShoppingLists.Update;
 
 public record ShoppingListUpdateDto(string Name, List<ShoppingSublistUpdateDto> Sublists);
 
-public record ShoppingSublistUpdateDto(int Id, int Count, List<ShoppingListItemUpdateDto> Items);
+public record ShoppingSublistUpdateDto(int Id, decimal Count, List<ShoppingListItemUpdateDto> Items);
 
 public record ShoppingListItemUpdateDto(int? Id, string Name, AmountDto Amount, bool Checked);
